requirements.txt
streamlit==1.33.0
google-generativeai==0.3.2
python-dotenv==1.0.1
youtube-transcript-api==0.6.2
beautifulsoup4==4.12.3
requests==2.31.0
langdetect==1.0.9
gTTS==2.4.0
sqlite3==2.6.0 
hashlib==20081119.1  
base64==1.0.1